random benches do strings of allocs and deallocs
Before, only one malloc or free was performed
every iteration. Now a random number of them are
performed to reflect that allocations often happen
multiples of times before the frees.

When doing only one at a time, it was testing
speed of allocations on a relatively unused heap.
Hopefully this change reflects a larger range
of loads on the heap.